Pull graph logic out into a graph helper class
